Updated architecture, separated type converters into their own .h/.cpp
