+++ TestCpp/source/Test.cpp
@@ -1,10 +1,10 @@
 
-#include "MathWorks.h"
+#include "C_Coder.h"
 
 void TestDemosaic()
 {
 	const cv::Mat &BayerImage = cv::imread("BayerPattern.png", cv::IMREAD_ANYDEPTH | cv::IMREAD_ANYCOLOR);
-	const cv::Mat &ColorImage = MathWorks::Functions::Demosaic(BayerImage, MathWorks::SensorAlignment::rggb);
+	const cv::Mat &ColorImage = MathWorks::C_Coder::Demosaic(BayerImage, MathWorks::SensorAlignment::rggb);
 
 	cv::imwrite("Color.png", ColorImage);
 }
@@ -12,9 +12,9 @@ void TestDemosaic()
 int main()
 {
 
-	const cv::Mat Image48Bit = cv::imread("24bit.png", cv::IMREAD_ANYDEPTH | cv::IMREAD_ANYCOLOR);
+	const cv::Mat Image48Bit = cv::imread("48bit.png", cv::IMREAD_ANYDEPTH | cv::IMREAD_ANYCOLOR);
 
-	const cv::Mat ResizedImage = MathWorks::Functions::Imresize(Image48Bit, 1000, 1000, MathWorks::ResizeMode::billinear);
+	const cv::Mat ResizedImage = MathWorks::C_Coder::Imresize(Image48Bit, 1000, 1000, MathWorks::ResizeMode::billinear);
 
 	std::cout << cv::imwrite("ResizedImage.png", ResizedImage) << "\n";
